rename dpgo and refactor UTF16 to winutil pkg
