+++ pkg/disk/dpgo/createDisk.TEMPLATE.txt
