+++ pkg/disk/dpgo/diskpart_windows.go
