+++ pkg/winutil/io.go
@@ -1,0 +1,26 @@
+package winutil
+
+import (
+	"io"
+
+	"golang.org/x/text/encoding/unicode"
+	"golang.org/x/text/transform"
+)
+
+// FromUTF16le returns an io.Reader for UTF16le data.
+// Windows uses little endian by default, use unicode.UseBOM policy to retrieve BOM from the text,
+// and unicode.LittleEndian as a fallback.
+func FromUTF16le(r io.Reader) io.Reader {
+	o := transform.NewReader(r, unicode.UTF16(unicode.LittleEndian, unicode.UseBOM).NewDecoder())
+	return o
+}
+
+// FromUTF16leToString reads from Unicode 16 LE encoded data from an io.Reader and returns a string.
+func FromUTF16leToString(r io.Reader) (string, error) {
+	out, err := io.ReadAll(FromUTF16le(r))
+	if err != nil {
+		return "", err
+	}
+
+	return string(out), nil
+}
